Fix a bug in node expansion (#66)
When a thread tries to expand the same node another thread is already expanding, the thread has to wait for the other thread to realease to lock on actions.
However after the lock is released, the new edges are already added, meaning the second thread will duplicate the edges.
The thing I suspect to be the reason this is super bad is because the first thread set all his Node indices to NodeIndex::NULL, which is a large number.
The second thread now reads those large numbers when trying to calculate the policy total, leading to the total being inf (exp(u32::MAX / 1000) == inf, its not actually u32::MAX / 1000 because of floating points tuff but quite close).
This then leads to the second thread relabeling all policies to zero (p / inf == 0) meaning the edges are selected soley based on q in the expansion phase.
This is sometimes quite bad leading to blunders.

Score of Jackal-3t vs Jackal-1t: 108 - 31 - 41  [0.714] 180
...      Jackal-3t playing White: 77 - 2 - 11  [0.917] 90
...      Jackal-3t playing Black: 31 - 29 - 30  [0.511] 90
...      White vs Black: 106 - 33 - 41  [0.703] 180
Elo difference: 158.8 +/- 48.1, LOS: 100.0 %, DrawRatio: 22.8 %

Score of Jackal-4t vs Jackal-2t: 93 - 53 - 38  [0.609] 184
...      Jackal-4t playing White: 70 - 10 - 13  [0.823] 93
...      Jackal-4t playing Black: 23 - 43 - 25  [0.390] 91
...      White vs Black: 113 - 33 - 38  [0.717] 184
Elo difference: 76.8 +/- 45.8, LOS: 100.0 %, DrawRatio: 20.7 %